+++ packages/core/src/controllers/ConnectorController.ts
@@ -1,6 +1,6 @@
 import { subscribeKey as subKey } from 'valtio/vanilla/utils'
 import { proxy, ref, snapshot } from 'valtio/vanilla'
-import type { Connector, EmailConnector } from '../utils/TypeUtil.js'
+import type { AuthConnector, Connector } from '../utils/TypeUtil.js'
 import { getW3mThemeVariables } from '@web3modal/common'
 import { OptionsController } from './OptionsController.js'
 import { ThemeController } from './ThemeController.js'
@@ -29,20 +29,21 @@ export const ConnectorController = {
     state.connectors = connectors.map(c => ref(c))
   },
 
-  addConnector(connector: Connector) {
+  addConnector(connector: Connector | AuthConnector) {
     state.connectors.push(ref(connector))
 
-    if (connector.id === 'w3mEmail') {
-      const emailConnector = connector as EmailConnector
+    if (connector.id === 'w3mAuth') {
+      const authConnector = connector as AuthConnector
       const optionsState = snapshot(OptionsController.state) as typeof OptionsController.state
       const themeMode = ThemeController.getSnapshot().themeMode
       const themeVariables = ThemeController.getSnapshot().themeVariables
-      emailConnector?.provider?.syncDappData?.({
+
+      authConnector?.provider?.syncDappData?.({
         metadata: optionsState.metadata,
         sdkVersion: optionsState.sdkVersion,
         projectId: optionsState.projectId
       })
-      emailConnector.provider.syncTheme({
+      authConnector.provider.syncTheme({
         themeMode,
         themeVariables,
         w3mThemeVariables: getW3mThemeVariables(themeVariables, themeMode)
@@ -50,8 +51,8 @@ export const ConnectorController = {
     }
   },
 
-  getEmailConnector() {
-    return state.connectors.find(c => c.type === 'EMAIL') as EmailConnector | undefined
+  getAuthConnector() {
+    return state.connectors.find(c => c.type === 'AUTH') as AuthConnector | undefined
   },
 
   getAnnouncedConnectorRdns() {