+++ packages/scaffold/src/partials/w3m-header/index.ts
@@ -103,13 +103,12 @@ export class W3mHeader extends LitElement {
           data-testid="w3m-header-close"
         ></wui-icon-link>
       </wui-flex>
-      ${this.separatorTemplate()}
     `
   }
 
   // -- Private ------------------------------------------- //
 
-  // Tempory added to test connecting with SIWE, replace with 'WhatIsAWallet' again when approved
+  // Temporarily added to test connecting with SIWE, replace with 'WhatIsAWallet' again when approved
   private onWalletHelp() {
     EventsController.sendEvent({ type: 'track', event: 'CLICK_WALLET_HELP' })
     RouterController.push('WhatIsAWallet')
@@ -134,8 +133,9 @@ export class W3mHeader extends LitElement {
     const isConnectHelp = view === 'Connect'
     const isApproveTransaction = view === 'ApproveTransaction'
     const isUpgradeToSmartAccounts = view === 'UpgradeToSmartAccount'
+    const isConnectingSIWEView = view === 'ConnectingSiwe'
 
-    const shouldHideBack = isApproveTransaction || isUpgradeToSmartAccounts
+    const shouldHideBack = isApproveTransaction || isUpgradeToSmartAccounts || isConnectingSIWEView
 
     if (this.showBack && !shouldHideBack) {
       return html`<wui-icon-link
@@ -152,14 +152,6 @@ export class W3mHeader extends LitElement {
       icon="helpCircle"
       @click=${this.onWalletHelp.bind(this)}
     ></wui-icon-link>`
-  }
-
-  private separatorTemplate() {
-    if (!this.heading) {
-      return null
-    }
-
-    return html`<wui-separator></wui-separator>`
   }
 
   private getPadding() {
@@ -219,11 +211,7 @@ export class W3mHeader extends LitElement {
   }
 
   private onGoBack() {
-    if (RouterController.state.view === 'ConnectingSiwe') {
-      RouterController.push('Connect')
-    } else {
-      RouterController.goBack()
-    }
+    RouterController.goBack()
   }
 }