+++ packages/scaffold/src/views/w3m-onramp-tokens-select-view/index.ts
@@ -38,7 +38,7 @@ export class W3mOnrampTokensView extends LitElement {
   // -- Render -------------------------------------------- //
   public override render() {
     return html`
-      <wui-flex flexDirection="column" padding="s" gap="xs">
+      <wui-flex flexDirection="column" .padding=${['0', 's', 's', 's']} gap="xs">
         ${this.currenciesTemplate()}
       </wui-flex>
       <w3m-legal-footer></w3m-legal-footer>