+++ packages/ui/src/composites/wui-details-group/styles.ts
@@ -4,7 +4,7 @@ export default css`
   :host {
     display: block;
     padding: var(--wui-spacing-l) var(--wui-spacing-m);
-    background-color: var(--wui-gray-glass-002);
+    background-color: var(--wui-color-gray-glass-002);
     border-radius: var(--wui-border-radius-xs);
     width: 100%;
   }