+++ packages/wagmi/src/utils/defaultWagmiReactConfig.ts
@@ -2,11 +2,10 @@ import '@web3modal/polyfills'
 
 import type { CreateConfigParameters, CreateConnectorFn, Config } from 'wagmi'
 import { createConfig } from 'wagmi'
-import { coinbaseWallet } from 'wagmi/connectors'
-
-import { emailConnector } from '../connectors/EmailConnector.js'
-import { alphaWalletConnect } from '../connectors/alphaWalletConnect.js'
+import { coinbaseWallet, walletConnect, injected } from 'wagmi/connectors'
+import { authConnector } from '../connectors/AuthConnector.js'
 import { getTransport } from './helpers.js'
+import type { SocialProvider } from '@web3modal/scaffold-utils'
 
 export type ConfigOptions = Partial<CreateConfigParameters> & {
   chains: CreateConfigParameters['chains']
@@ -14,10 +13,9 @@ export type ConfigOptions = Partial<CreateConfigParameters> & {
   enableEIP6963?: boolean
   enableCoinbase?: boolean
   enableEmail?: boolean
-  /**
-   * Use enableEIP6963 to show all injected wallets
-   * @deprecated
-   */
+  auth?: {
+    socials?: SocialProvider[]
+  }
   enableInjected?: boolean
   enableWalletConnect?: boolean
   metadata: {
@@ -34,8 +32,9 @@ export function defaultWagmiConfig({
   metadata,
   enableCoinbase,
   enableEmail,
+  enableInjected,
+  auth,
   enableWalletConnect,
-
   enableEIP6963,
   ...wagmiConfig
 }: ConfigOptions): Config {
@@ -48,7 +47,11 @@ export function defaultWagmiConfig({
 
   // Enabled by default
   if (enableWalletConnect !== false) {
-    connectors.push(alphaWalletConnect({ projectId, metadata, showQrModal: false }))
+    connectors.push(walletConnect({ projectId, metadata, showQrModal: false }))
+  }
+
+  if (enableInjected !== false) {
+    connectors.push(injected({ shimDisconnect: true }))
   }
 
   if (enableCoinbase !== false) {
@@ -62,8 +65,15 @@ export function defaultWagmiConfig({
   }
 
   // Dissabled by default
-  if (enableEmail === true) {
-    connectors.push(emailConnector({ chains: [...chains], options: { projectId } }))
+  if (enableEmail || auth?.socials) {
+    connectors.push(
+      authConnector({
+        chains: [...chains],
+        options: { projectId },
+        socials: auth?.socials,
+        email: enableEmail
+      })
+    )
   }
 
   return createConfig({